Complete version without motors
This push implements the full code excluding the motor control.

+++ firmware/include/states.h
@@ -18,9 +18,9 @@ void handleRefill(FSM* nFsm);
 
 //Helper functions
 char classifyCup(float weight);
-uint8_t calculateMl(FSM* nFsm, uint8_t pump);
+uint16_t calculateMl(FSM* nFsm, uint8_t pump);
 uint16_t getCupSize(FSM* nFsm);
 void dynamicDelay (FSM* nFsm, uint8_t pump);
 void rstAction(FSM* nFsm);
 void potClassify();
-#endif // !DEBUG
+#endif // !STATES
