+++ firmware/include/timers.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-uint16_t calculateTime(uint8_t volume);
+uint16_t calculateTime(uint16_t volume);
 uint16_t calculateOCR1(uint16_t time);
 void dynamicTimer(uint16_t ocr);
 void delay5s();
